+++ js/const.js
@@ -1,0 +1,38 @@
+const DESCRIPTIONS = [
+  'Закат над морем',
+  'Улыбающийся ребёнок',
+  'Лес осенью',
+  'Городские огни',
+  'Цветущий сад',
+  'Снежные вершины',
+  'Пляжный отдых',
+  'Ночной город',
+  'Весенние цветы',
+  'Пустынный пейзаж'
+];
+
+const NAMES = [
+  'Анна',
+  'Дмитрий',
+  'Екатерина',
+  'Иван',
+  'Мария',
+  'Николай',
+  'Ольга',
+  'Павел',
+  'Светлана',
+  'Татьяна'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const PHOTOS_COUNT = 25;
+
+export { DESCRIPTIONS, NAMES, COMMENTS, PHOTOS_COUNT };
